refactor: split to view, controller, domain package
