+++ src/main/java/nextstep/ladder/Main.java
@@ -1,0 +1,7 @@
+package nextstep.ladder;
+
+public class Main {
+    public static void main(String[] args) {
+        
+    }
+}
