+++ src/main/java/nextstep/ladder/domain/Connection.java
@@ -1,4 +1,4 @@
-package nextstep.ladder;
+package nextstep.ladder.domain;
 
 import java.util.Objects;
