+++ src/main/java/nextstep/ladder/domain/ConnectionCreationStrategy.java
@@ -1,4 +1,4 @@
-package nextstep.ladder;
+package nextstep.ladder.domain;
 
 @FunctionalInterface
 interface ConnectionCreationStrategy {
