+++ src/main/java/nextstep/ladder/domain/Players.java
@@ -1,4 +1,4 @@
-package nextstep.ladder;
+package nextstep.ladder.domain;
 
 import java.util.Arrays;
 import java.util.List;
